+++ src/main/java/racingcar/domain/Participants.java
@@ -5,7 +5,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.utils.NumberGenerator;
 
 public class Participants {
