+++ src/test/java/racingcar/domain/CarTest.java
@@ -3,8 +3,6 @@ package racingcar.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import racingcar.utils.DrivableNumberGenerator;
-import racingcar.utils.NonDrivableNumberGenerator;
 
 class CarTest {
