+++ src/test/java/racingcar/domain/DrivableNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.domain;
 
 public class DrivableNumberGenerator implements NumberGenerator {
