+++ src/test/java/racingcar/domain/RaceTest.java
@@ -8,8 +8,6 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.utils.NumberGenerator;
-import racingcar.utils.RandomNumberGenerator;
 
 class RaceTest {
